move 6 to root level
