+++ src/routes/(site)/hero.svelte
@@ -2,76 +2,97 @@
   import AcmButton from '$lib/components/button/button.svelte';
 </script>
 
-<section>
-  <div class="container">
-    <div class="left">
+<section class="hero-container">
+  <div class="hero-inner-container">
+    <div class="hero-text">
       <h1 class="brand-header size-xl">
-        We are the largest computer science community at
-        <b class="acm-blue">CSUF</b>
+        <span>We are the largest</span> <span class="multicolor-text">Computer Science</span>
+        <span>community at CSUF</span>
       </h1>
-
-      <AcmButton text="Join us today!" link="/discord" redirect={true} />
+      <div class="join-button">
+        <AcmButton text="Join today!" link="/discord" redirect={true} />
+      </div>
     </div>
 
-    <img
-      src="assets/hero-photo.png"
-      alt="Chip the Capybara holding a flag that says 'I Love ACM'"
-    />
+    <img src="/assets/capy-power.svg" alt="Chip the Capybara surrounded by all ACM Team badges" />
   </div>
 </section>
 
-<style lang="scss">
+<style>
   section {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    .container {
-      display: flex;
-      align-items: center;
-      justify-content: space-between;
-      padding: 0 16px;
-      width: 1280px;
-
-      .left {
-        display: flex;
-        align-items: flex-start;
-        flex-direction: column;
-        gap: 64px;
-
-        h1 {
-          max-width: clamp(21.875rem, 18.8048rem + 17.5439vw, 46.875rem);
-        }
-      }
-
-      img {
-        width: clamp(20rem, 17.342rem + 10.13vw, 30rem);
-        margin-left: 32px;
-        filter: drop-shadow(0 10px 40px rgba(44, 145, 198, 0.5));
-      }
-    }
+    display: grid;
   }
 
-  @media screen and (max-width: 900px) {
-    section .container {
-      padding: 0 8px;
-      flex-direction: column-reverse;
+  section .hero-inner-container {
+    display: grid;
+    grid-template-columns: 1fr;
+    align-items: center;
+    gap: 1em;
+  }
 
-      .left {
-        align-items: center;
-        text-align: center;
-        gap: 32px;
+  section .hero-inner-container .hero-text {
+    display: grid;
+    align-items: center;
+    text-align: center;
+    gap: 1em;
+  }
 
-        h1 {
-          font-size: var(--size-lg);
-        }
-      }
+  section .hero-inner-container .hero-text span {
+    display: block;
+  }
 
-      img {
-        margin: 0;
-        padding-bottom: 64px;
-        filter: drop-shadow(0 6px 24px rgba(44, 145, 198, 0.5));
-      }
+  /* Since we gave a display block to the span elements within the h1, 
+   * each will cover the entire page width, causing issues with the 
+   * multicolor linear-gradient (covering the webpage width rather than 
+   * the actual text ONLY). Giving a display of inline-block to the span 
+   * containing the multicolor text resolved this issue.
+   */
+  section .hero-inner-container .hero-text span.multicolor-text {
+    display: inline-block;
+    background: linear-gradient(
+      to right,
+      var(--acm-turquoise),
+      var(--acm-blue),
+      var(--acm-bluer),
+      var(--acm-purple),
+      var(--acm-pink)
+    );
+    background-clip: text;
+    -webkit-background-clip: text;
+    color: transparent;
+  }
+
+  section .hero-inner-container .hero-text .join-button {
+    display: grid;
+    justify-self: center;
+  }
+
+  section .hero-inner-container img {
+    max-width: 90%;
+    justify-self: center;
+  }
+
+  @media screen and (min-width: 1000px) {
+    section {
+      display: grid;
+      place-items: center;
+      align-items: center;
+    }
+
+    section .hero-inner-container {
+      display: grid;
+      grid-template-columns: 1fr 1fr;
+      align-items: center;
+      gap: 4em;
+    }
+
+    section .hero-inner-container .hero-text {
+      text-align: start;
+      margin: 2em;
+    }
+
+    section .hero-inner-container .hero-text .join-button {
+      justify-self: start;
     }
   }
 </style>
