+++ src/routes/(site)/lucky.svelte
