+++ src/routes/(site)/teams/diamond-picture.svelte
